+++ lib/index.ts
@@ -2,4 +2,5 @@ export * from './module';
 export * from './page';
 export * from './pib';
 export * from './provider';
+export * from './routing';
 export * from './storage';
